requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
mysql-connector-python==9.0.0
google-generativeai==0.7.2
gunicorn==23.0.0
